Move TriState to utils package
