--- Essentials/src/main/java/com/earth2me/essentials/perm/TriState.java
@@ -1,10 +1,0 @@
-package com.earth2me.essentials.perm;
-
-/**
- * The possible "exact" values for a permission node lookup.
- */
-public enum TriState {
-    TRUE,
-    FALSE,
-    UNSET
-}
